Just storing my WIP changes for publicity

+++ pkg/avail/client.go
@@ -17,7 +17,7 @@ type Client interface {
 	BlockStream(offset uint64) BlockStream
 	GenesisHash() types.Hash
 	GetLatestHeader() (*types.Header, error)
-	SearchBlock(offset int, searchFunc SearchFunc) (*types.SignedBlock, error)
+	SearchBlock(offset int64, searchFunc SearchFunc) (*types.SignedBlock, error)
 }
 
 type client struct {
